Hint 중국어 번역 수정 중국어 번역 추가 data type casting 추가

+++ JMCClasses/Core/inbox/JMCIssueStore.m
@@ -120,7 +120,7 @@ static NSRecursiveLock *writeLock;
                                    "dateCreated, "
                                    "hasUpdates "
                                 "FROM issue ORDER BY hasUpdates desc, dateUpdated desc LIMIT 1 OFFSET ?",
-                           [NSNumber numberWithUnsignedInt:issueIndex]];
+                           [NSNumber numberWithUnsignedInt:(unsigned int)issueIndex]];
     if ([res next]) {
         NSDictionary* dictionary = [res resultDict];
         JMCIssue* issue = [[JMCIssue alloc] initWithDictionary:dictionary];
@@ -130,7 +130,7 @@ static NSRecursiveLock *writeLock;
         }
         return issue;
     }
-    JMCALog(@"No issue at index = %u", issueIndex);
+    JMCALog(@"No issue at index = %u", (unsigned int)issueIndex);
     return nil;
 }
